refactor: move AnimatedListfootercomponent to its own file
